board write error fixed commit

+++ src/main/java/alz/board/controller/BoardPageController.java
@@ -90,7 +90,7 @@ public class BoardPageController {
 	}
 	
 	@GetMapping("/write")
-	public void write(@RequestParam("typeId") Long typeId, Model model) {
+	public void write(@RequestParam("typeId") Integer typeId, Model model) {
 		model.addAttribute("typeId", typeId);
 	}
